web/python/requirements.txt
openai-whisper>=20231117
torch>=2.3.0
numpy>=1.26.0
